+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # PrivateBin version history
 
+  * **next (not yet released)**
+    * ADDED: Translations for Spanish
   * **1.1 (2016-12-26)**
     * ADDED: Translations for Italian and Russian
     * ADDED: Loading message displayed until decryption succeeded for slower (in terms of CPU or network) systems